+++ src/test/java/io/geekidea/boot/generator/Generator.java
@@ -1,6 +1,7 @@
 package io.geekidea.boot.generator;
 
 import io.geekidea.boot.generator.config.GeneratorConfig;
+import io.geekidea.boot.generator.enums.RequestMappingType;
 import io.geekidea.boot.generator.handler.GenerateHandler;
 
 /**
@@ -35,6 +36,18 @@ public class Generator {
         config.setQueryFileName("Request");
         config.setInfoVoFileName("DetailVo");
         config.setVoFileName("Response");
+        // 自定义请求路径
+        config.setControllerRequestMapping("%s");
+        config.setAddRequestMapping("add%s");
+        config.setUpdateRequestMapping("update%s");
+        config.setDeleteRequestMapping("delete%s");
+        config.setInfoRequestMapping("get%sInfo");
+        config.setPageRequestMapping("get%sPage");
+        // 自定义请求路径类型 CAMEL：骆驼(默认)，HYPHEN：中横线，下划线：UNDERLINE，反斜杠：BACKSLASH，全小写：LOWER
+        config.setRequestMappingType(RequestMappingType.LOWER);
+        config.setRequestMappingModule(true);
+
+
 
         GenerateHandler handler = new GenerateHandler();
         handler.generator(config);