+++ src/test/java/io/geekidea/boot/generator/enums/RequestMappingType.java
@@ -1,0 +1,31 @@
+package io.geekidea.boot.generator.enums;
+
+/**
+ * @author geekidea
+ * @date 2022/7/3
+ **/
+public enum RequestMappingType {
+
+    /**
+     * 骆驼命名法
+     */
+    CAMEL,
+    /**
+     * 中横线连接
+     */
+    HYPHEN,
+    /**
+     * 下划线
+     */
+    UNDERLINE,
+    /**
+     * 反斜杠
+     */
+    BACKSLASH,
+    /**
+     * 全小写
+     */
+    LOWER;
+
+
+}